nvme support (wip)
This is the logic I'm currently using to try and install to an nvme disk. Unfortunately still not working correctly. I made some hardcoded changes for loading the EFI bootloader adding the letter p but for some odd reason, it isn't reflecting this. During post-install tasks, it errors out installing the bootloader to /dev/nvme0n12 when it should be /dev/nvme0n1p2, I don't know why the letter p isn't being added on.