Don't shrink disk selection dialog
After the disk selection dialog message was changed, we took up another
line and squeezed the list down smaller than it needs to be.  The math
here was full of magic numbers to attempt to shrink the list area to fit
the disk list exactly if under a certain size, which makes changing it
very cumbersome.

Simplify things by always using the max dialog size and letting dialog
determine the available space for the list area automatically.