Fix check for mounted disks, simplify disk listing
Update the partition regex for Linux in disk_is_mounted so we can use it
to filter mounted disks out of the disk selection menu list.  This
avoids offering to install over the installer media.

While here, also filter out names that do not correspond to an actual
block device and disks that are not big enough.

There was some crude filtering based on name prefixes to avoid some
vague issue about MMC block device regions and something about NVMe but
the logic behind it was unclear. The new checks should be sufficient, so
simplify things by removing the old filtering mechanism.

Jira: NAS-111558